Slice rotations over x,y,z
Middle layer rotations are normally called M (middle), E (edge), S
(slice).
M: rotation over x
E: rotation over y
S: rotation over z
Slice rotations will also have buttons on the page for manual control.

This definition of x,y,z is quite arbitrary and doesn't reflect common
mathematical views.

The problem with implementing these as opposed to side face rotations,
is that the center of the layer moves with the rotation. That means the
view of what is 'up', or 'right' gets confused. You can choose to either
keep a 'camera' view of the cube, and decide Right is whatever looks to
be on the right for the camera, or you choose that since the center of
the Right face has moved, Right itself is now e.g. 'in the back'. The
latter is a more rigid definition that helps with performing algorithms,
but it's not helpful for someone looking at the cube on screen.

I have chosen to keep Right, Up and Front as they are viewed on screen.
For that we need to implement a separate centerCubes array to find out
what layer is where.

Closes #15