Initial algorithm
There is much more to be done for parameters. But this is a good start,
so I am closing it. Next I will work on algorithm stuff like 'generator'
vs 'solver'.

The algorithm can either be set using a data attribute in the canvas
element. So e.g. data-alg="RLFBUD", or you can enter it in the url as
?alg=...

Development will now continue on index.php instead of .html

Closes #22