More realistic perspective view
This looks definitely better. However, I have returned more to the
original view of the cube, which means the 'eye' is closer. This gives
more perspective effect to the cube. This could be either annoying, or a
great effect. What kind of cube do we want? A realistic one, a practical
one or a beautiful one? Those are not entirely mutually exclusive.

Closes #45